Update the Readme forking notes
The platform does not support `main` for custom upstreams. While this can be changed by Pantheors in the admin dashboard, that is not possible and adds a point of friction to users. Instead, as much as it pains us, we need to ask users to create a `master` branch and set that as the default for their custom upstreams, at least for the time being.