Update to resolve repository maintenance issues
This update to the WordPress (Composer Managed) upstream disables
Dependabot dependency management notifications (we do not want to change
the versions of dependencies in our composer.json file for risk of merge
conflicts).

This update also resolves an issue that was causing releases to not
deploy to the Decoupled WordPress (Composer Managed) upstream.