Roots Sage install script & repository maintenance
This update adds an install script (located in /private/scripts and accessible via Composer) that can assist in installing the Roots Sage theme on new WordPress (Composer Managed) sites. Documentation can be found in the README.md file and in the Installing Sage doc in /docs.  This update also disables the Dependabot dependency management notifications (we do not want to change the versions of dependencies in our composer.json file for risk of merge conflicts).  Finally, this update resolves an issue that was causing releases to not deploy to the Decoupled WordPress (Composer Managed) upstream.