[BUGS-8491] only set `DISABLE_WP_CRON` value once (#139)
* remove wp-config-pantheon require

* remove DISABLE_WP_CRON constant from application.php
this should be set in application.pantheon.php not in application.php. application.php can override the value set in application.pantheon.php if the constant is defined before application.pantheon.php is loaded

* pull out everything in wp-config-pantheon
application.pantheon.php is the place for these things now

* set the temp dir (from wp-config-pantheon)

* set WP_ENVIRONMENT_TYPE (from wp-config-pantheon)

* set DISABLE_WP_CRON (from wp-config-pantheon)
This is slightly altered from how it existed in wp-config-pantheon in that it's possible to override. It does not define if it's already defined (meaning you can define it yourself) but it's not recommended

* use tabs

* linting fixes

* delete wp-config-pantheon