[SITE-3292] add @maybe-add-symlinks to post-update-cmd (#175)
* add @maybe-add-symlinks to post-update-cmd
add the symlinks if they don't already exist

* Add symlinks on composer update command as well as composer install.

This update applies the same check to create symlinks to files in the `web/` directory that is run on `composer install` commands to `composer update` as well. This ensures that symlinks are created with upstream updates are applied (`composer update`) as well as when new commits to a site repository are pushed (`composer install`).

For more information, see https://docs.pantheon.io/release-notes/2025/02/wordpress-composer-managed-1-32-5